Improve filter fragment embedding
Handle the case where we add a filter to an existing query, so that
now it works even if there was no filter to start with.  Since a
missing filter is equivalent to a filter that accepts everything,
adding a filter with 'and' is equivalent to just using the new filter,
while 'or' and 'not' don't make any sense with an empty filter.  Let's
raise errors as appropriate.

Also use the caret operator when embedding filter fragments, mimicking
Ecto's "bindingless" style.