Delay encoding nested queries
When using an inner query or a join as data source in the `from`
macro, don't encode it to a JSON-like map immediately.  Instead, do it
when we encode the outer query.

This should make it easier to modify the inner query as well as the
outer query before encoding.